split main to two files
