+++ src/log.rs
@@ -1,4 +1,4 @@
-pub const DEBUG: bool = false;
+pub const DEBUG: bool = true;
 
 macro_rules! debug {
 	($($arg:tt)*) => {
